Implement key derivation at the storage engine
- Used Key.derivePBKDF2SHA256Key() method implemented in couchbase-lite-java-sqlcipher to do key deriviation with libcrypto.

- Moved all SymmetricKey tests from MiscTest into DatabaseEncryptionTest.

- Moved ActionTest into support folder

https://github.com/couchbase/couchbase-lite-java-core/issues/805
